+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - cURL output format (rif <filename> --output=curl)
 ### Changed
  - Improved error message when a variable is given an unknown type in a RIF file
  - Improved error message when required variables are not given