+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.4.3] - 2018-11-19
+### Changed
+ - Various refactorings
+
 ## [0.4.2] - 2018-04-11
 ### Changed
  - Improved error message when rif\_version field is missing